+++ pkg/contexts/credentials/builtin/npm/identity/identity.go
@@ -1,0 +1,63 @@
+package identity
+
+import (
+	"path"
+
+	. "net/url"
+
+	"github.com/open-component-model/ocm/pkg/common"
+	"github.com/open-component-model/ocm/pkg/contexts/credentials/cpi"
+	"github.com/open-component-model/ocm/pkg/contexts/credentials/identity/hostpath"
+	"github.com/open-component-model/ocm/pkg/listformat"
+)
+
+const (
+	// CONSUMER_TYPE is the npm repository type.
+	CONSUMER_TYPE = "Registry.npm.com"
+
+	// ATTR_USERNAME is the username attribute. Required for login at any npm registry.
+	ATTR_USERNAME = cpi.ATTR_USERNAME
+	// ATTR_PASSWORD is the password attribute. Required for login at any npm registry.
+	ATTR_PASSWORD = cpi.ATTR_PASSWORD
+	// ATTR_EMAIL is the email attribute. Required for login at any npm registry.
+	ATTR_EMAIL = cpi.ATTR_EMAIL
+	// ATTR_TOKEN is the token attribute. May exist after login at any npm registry.
+	ATTR_TOKEN = cpi.ATTR_TOKEN
+)
+
+func init() {
+	attrs := listformat.FormatListElements("", listformat.StringElementDescriptionList{
+		ATTR_USERNAME, "the basic auth user name",
+		ATTR_PASSWORD, "the basic auth password",
+		ATTR_EMAIL, "NPM registry, require an email address",
+		ATTR_TOKEN, "the token attribute. May exist after login at any npm registry. Check your .npmrc file!",
+	})
+
+	cpi.RegisterStandardIdentity(CONSUMER_TYPE, hostpath.IdentityMatcher(CONSUMER_TYPE), `NPM repository
+
+It matches the <code>`+CONSUMER_TYPE+`</code> consumer type and additionally acts like 
+the <code>`+hostpath.IDENTITY_TYPE+`</code> type.`,
+		attrs)
+}
+
+func GetConsumerId(rawURL string, pkgName string) cpi.ConsumerIdentity {
+	url, err := Parse(rawURL)
+	if err != nil {
+		return nil
+	}
+
+	url.Path = path.Join(url.Path, pkgName)
+	return hostpath.GetConsumerIdentity(CONSUMER_TYPE, url.String())
+}
+
+func GetCredentials(ctx cpi.ContextProvider, repoUrl string, pkgName string) common.Properties {
+	id := GetConsumerId(repoUrl, pkgName)
+	if id == nil {
+		return nil
+	}
+	credentials, err := cpi.CredentialsForConsumer(ctx.CredentialsContext(), id)
+	if credentials == nil || err != nil {
+		return nil
+	}
+	return credentials.Properties()
+}
